Revert "refactor and add types"
This reverts commit b4c552fa
