+++ tests/trainer/test_dataloaders.py
@@ -2,7 +2,7 @@ import pytest
 
 import tests.base.utils as tutils
 from pytorch_lightning import Trainer
-from pytorch_lightning.debugging import MisconfigurationException
+from pytorch_lightning.utilities.exceptions import MisconfigurationException
 from tests.base import (
     TestModelBase,
     LightningTestModel,
